trying with jaccard 0.5

+++ src/bundle_bridge.h
@@ -77,7 +77,7 @@ public:
 	int min(int x, int y, int z);
 	int get_edit_distance(string s, string t);
 	bool are_strings_similar(int kmer_length, map<string,int> kmer_map, string t);
-	float get_Jaccard(int kmer_length, map<string,int> kmer_map, string t);
+	double get_Jaccard(int kmer_length, map<string,int> kmer_map, string t);
 	int get_more_chimeric();
 	int create_fake_fragments();
 	int create_fake_supple(int fr_index, fragment &fr, int32_t soft_len, int32_t pos1, int32_t pos2, int soft_clip_side);
